+++ ISHPermissionKit/Private/ISHPermissionRequestMusicLibrary.h
@@ -1,0 +1,18 @@
+//
+//  ISHPermissionRequestMusicLibrary.h
+//  ISHPermissionKit
+//
+//  Created by Hagi on 26/08/16.
+//  Copyright © 2016 iosphere GmbH. All rights reserved.
+//
+
+#import <ISHPermissionKit/ISHPermissionKit.h>
+
+/**
+ *  Permission request to access the user's music library.
+ *
+ *  @sa ISHPermissionCategoryMusicLibrary
+ */
+@interface ISHPermissionRequestMusicLibrary : ISHPermissionRequest
+
+@end